refactor: organise json_schema_builder JS logic into folders
