+++ nodes/web/helpers/json_schema_builder/main.js
@@ -1,0 +1,3 @@
+import { handleNumSlots, widgetLogic } from "./utils.js";
+
+export { handleNumSlots, widgetLogic };
