+++ nodes/web/json_schema_builder.js
@@ -1,46 +1,5 @@
 import { app } from "../../../scripts/app.js";
-
-function handleNumSlots(node, widget) {
-  const numSlots = parseInt(widget.value);
-
-  if (!node.widgetsHidden) {
-    node.widgetsHidden = [];
-  }
-
-  for (let i = 1; i <= 10; i++) {
-    const widgets = [...node.widgets, ...node.widgetsHidden].filter((w) => w.name.endsWith(`_${i}`));
-
-    widgets.forEach((w) => {
-      if (i <= numSlots) {
-        const hiddenIndex = node.widgetsHidden.indexOf(w);
-        const widgetIndex = node.widgets.indexOf(w);
-        if (hiddenIndex !== -1 && widgetIndex === -1) {
-          node.widgets.push(w);
-          node.widgetsHidden.splice(hiddenIndex, 1);
-        }
-      } else {
-        const widgetIndex = node.widgets.indexOf(w);
-        if (widgetIndex !== -1) {
-          node.widgetsHidden.push(w);
-          node.widgets.splice(widgetIndex, 1);
-        }
-      }
-    });
-  }
-
-  node.setSize(node.computeSize());
-}
-
-const nodeWidgetHandlers = {
-  num_fields: handleNumSlots,
-};
-
-function widgetLogic(node, widget, isInitial = false) {
-  const handler = nodeWidgetHandlers[widget.name];
-  if (handler && !isInitial) {
-    handler(node, widget);
-  }
-}
+import { handleNumSlots, widgetLogic } from "./helpers/json_schema_builder/main.js";
 
 const ext = {
   name: "signature.json_schema_builder",
